+++ src/parser/block/extension/placeholder.test.ts
@@ -1,21 +1,23 @@
 ﻿import { placeholder } from './placeholder';
 import { some } from '../../../combinator';
-import { inspect } from '../../../debug.test';
 
 describe('Unit: parser/block/extension/placeholder', () => {
   describe('placeholder', () => {
     const parser = some(placeholder);
 
-    it('warning', () => {
-      assert.deepStrictEqual(inspect(parser(' ~~~\n~~~')), undefined);
-      assert.deepStrictEqual(inspect(parser('~~~\n~~~')), [["<p><em>Invalid syntax: Extension syntax: ~~~.</em></p>"], '']);
-      assert.deepStrictEqual(inspect(parser('~~~\n~~~\n')), [["<p><em>Invalid syntax: Extension syntax: ~~~.</em></p>"], '']);
-      assert.deepStrictEqual(inspect(parser('~~~\n<"\n~~~')), [["<p><em>Invalid syntax: Extension syntax: ~~~.</em></p>"], ""]);
-      assert.deepStrictEqual(inspect(parser('~~~a\n~~~')), [["<p><em>Invalid syntax: Extension syntax: ~~~.</em></p>"], '']);
-      assert.deepStrictEqual(inspect(parser('~~~a \n~~~')), [["<p><em>Invalid syntax: Extension syntax: ~~~.</em></p>"], '']);
-      assert.deepStrictEqual(inspect(parser('~~~a b \n~~~')), [["<p><em>Invalid syntax: Extension syntax: ~~~.</em></p>"], '']);
-      assert.deepStrictEqual(inspect(parser('~~~ a\n~~~')), [["<p><em>Invalid syntax: Extension syntax: ~~~.</em></p>"], '']);
-      assert.deepStrictEqual(inspect(parser('~~~~\n~~~~')), [["<p><em>Invalid syntax: Extension syntax: ~~~.</em></p>"], '']);
+    it('invalid', () => {
+      assert(!parser(' ~~~\n~~~'));
+    });
+
+    it('valid', () => {
+      assert(parser('~~~\n~~~'));
+      assert(parser('~~~\n~~~\n'));
+      assert(parser('~~~\n<"\n~~~'));
+      assert(parser('~~~a\n~~~'));
+      assert(parser('~~~a \n~~~'));
+      assert(parser('~~~a b \n~~~'));
+      assert(parser('~~~ a\n~~~'));
+      assert(parser('~~~~\n~~~~'));
     });
 
   });