+++ src/parser/block/math.ts
@@ -7,10 +7,10 @@ import { html } from 'typed-dom';
 export const segment: MathParser = block(build(() => segment_));
 
 export const segment_: MathParser = block(match(
-  /^\$\$[^\n]*(\n(?:[^\n]*\n)*?)\$\$[^\S\n]*(?:\n|$)/,
+  /^\$\$([^\n]*)(\n(?:[^\n]*\n)*?)\$\$[^\S\n]*(?:\n|$)/,
   (_, rest) => [[], rest]), false);
 
 export const math: MathParser = block(rewrite(segment, match(
-  /^\$\$[^\n]*(\n(?:[^\n]*\n)*?)\$\$\s*$/,
-  ([, body], rest) =>
-    [[html('div', { class: 'math notranslate' }, `$$${body}$$`)], rest])));
+  /^\$\$([^\n]*)(\n(?:[^\n]*\n)*?)\$\$\s*$/,
+  ([, arg, body], rest) =>
+    [[html('div', { class: `math notranslate ${arg.trim() ? 'invalid' : ''}`.trim() }, `$$${body}$$`)], rest])));