+++ src/parser/block/paragraph/reference.ts
@@ -1,5 +1,5 @@
 ﻿import { ParagraphParser } from '../../block';
-import { union, match, validate, eval } from '../../../combinator';
+import { union, some, match, validate, eval } from '../../../combinator';
 import { line } from '../../source/line';
 import '../../source/unescapable';
 import { inline } from '../../inline';
@@ -13,6 +13,6 @@ export const reference: ParagraphParser.ReferenceParser = line(validate(
       ([ref, { length: level }], rest) =>
         [[html('a', { class: 'reference', rel: 'noopener', 'data-level': `${level}` }, ref.trim()), html('br')], rest]),
     () =>
-      [[...eval(inline(`*Invalid syntax: Reference syntax: Use lower-case alphanumeric characters in reference syntax.*`)), html('br')], ''],
+      [[html('span', { class: 'invalid' }, eval(some(inline)(`Invalid syntax: Reference syntax: Use lower-case alphanumeric characters in reference syntax.`))), html('br')], ''],
   ])
 ), true, true);