+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Reserve attribute schema of math syntax.
 - Allow empty URI in media syntax.
 - Allow leading and trailing whitespace around URI in link and media syntax.
+- Allow invalid arguments of syntax and mark as invalid.
 
 ## 0.72.0